Add initial implementation of logger
Implemented logger as a singleton in common/src/logger.

I think some things can be improved. Overall it does the job of
abstracting standard output and implements levels of logging.

I have mixed logging and "general" prints together, which I am not sure
this is standard practice. I think there can be improvements. Of note is
ofc the missing module name for the prints. There might also be a slight
inefficiency when calling log() >> "", as the check for severity level
is done twice.

I will create a ticket to follow-up on this but for this this'll do.

	modified:   common/include/CMakeLists.txt
	modified:   common/include/common.h
	modified:   common/src/CMakeLists.txt
	modified:   common/src/converters/converters.h
	new file:   common/src/logger/CMakeLists.txt
	new file:   common/src/logger/logger.cpp
	new file:   common/src/logger/logger.h
	modified:   common/test/test_common_include/CMakeLists.txt
	modified:   common/test/test_common_include/test_common_include.cpp
	modified:   main/CMakeLists.txt
	modified:   main/main.cpp

Signed-off-by: 6367f766 <78221071+6367f766@users.noreply.github.com>