Change newline to make it somewhat modular
The intention here was to make newline configurable or not. There are
instances for example in the evaluate --file command, at the end of a
binary sequence, the number should be displayed in the same line:

0110 0111 0000 0000 0000 0000 0000 0000 100 <--- note 100 is the value
and it's in the same line.

	modified:   common/src/converters/converters.h

Although this works... I think this needs more thought.

Signed-off-by: 6367f766 <78221071+6367f766@users.noreply.github.com>